requirements.txt
# ---------------------------------------------
# 📊 DATA PROCESSING & ANALYSIS
# ---------------------------------------------
pandas>=2.0.0          # Data manipulation and analysis — core for working with tabular data
numpy>=1.26.4          # Numerical operations, arrays, and matrix support (used by almost all ML libraries)

# ---------------------------------------------
# 🧠 MACHINE LEARNING
# ---------------------------------------------
scikit-learn>=1.3.0    # Classic ML models (regression, classification, clustering), preprocessing, model evaluation
xgboost==1.7.5
# ---------------------------------------------
# 📈 VISUALIZATION
# ---------------------------------------------
matplotlib>=3.8.0      # Base library for plotting graphs, charts, and visualizations
seaborn>=0.13.0        # High-level interface for drawing attractive and informative statistical graphics (built on matplotlib)

# ---------------------------------------------
# 📦 EXPERIMENT TRACKING & MODEL MANAGEMENT
# ---------------------------------------------
mlflow>=2.8.0          # End-to-end ML lifecycle tracking, including experiment tracking, model packaging, and deployment
boto3>=1.34.0          # AWS SDK for Python - required for SageMaker and S3 integration
sagemaker>=2.190.0     # AWS SageMaker Python SDK for ML pipeline orchestration

# ---------------------------------------------
# ✅ TESTING
# ---------------------------------------------
pytest==7.3.1          # Python testing framework — great for writing unit tests for ML pipelines, data validation, etc.

# ---------------------------------------------
# ⚡ API DEVELOPMENT (FOR MODEL SERVING)
# ---------------------------------------------
fastapi>=0.104.0       # Lightweight, high-performance web framework for serving ML models via REST APIs
uvicorn>=0.24.0        # ASGI server for running FastAPI apps — lightweight and fast
#
# ---------------------------------------------
# 🔄 WORKFLOW ORCHESTRATION
# ---------------------------------------------
apache-airflow>=2.7.0  # Workflow orchestration platform for ML pipelines

# ---------------------------------------------
#  MISC
# ---------------------------------------------
pyyaml>=6.0.1
joblib==1.3.1
setuptools==65.5.0
ipykernel==6.29.5


